pyproject.toml
[project]
name = "actualiza-portadas-de-vaniadev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.1",
    "pytonotion==0.0.2",
]
